+++ src/pyeuvtools/__init__.py
@@ -1,3 +1,3 @@
 """pyEUVTools package."""
 
-__version__ = "0.1.3"
+__version__ = "0.2.0"